add remote simulation settings implementation, variables are now serializable

+++ alchemist/alchemist-grid/src/main/java/it/unibo/alchemist/grid/simulation/Complexity.java
@@ -1,7 +1,6 @@
 package it.unibo.alchemist.grid.simulation;
 
 public interface Complexity {
-    public float getRamUsage();
-    public float getCpuUsage();
-
+    public double getRamUsage();
+    public double getCpuUsage();
 }
